apply `cargo +nightly clippy --fix`

+++ profiling/src/capi.rs
@@ -33,7 +33,7 @@ impl<'a> From<&'a [u8]> for StringView<'a> {
 impl<'a> StringView<'a> {
     pub fn as_slice(&self) -> &'a [u8] {
         // Safety: StringView's are required to uphold these invariants.
-        unsafe { core::slice::from_raw_parts(self.ptr as *const u8, self.len as usize) }
+        unsafe { core::slice::from_raw_parts(self.ptr as *const u8, self.len) }
     }
 
     pub fn to_string_lossy(&self) -> Cow<'a, str> {
